+++ src/third/ReactComponent.ts
@@ -59,8 +59,7 @@ export class ReactComponent extends SimpleComponent {
 
     async execute(block: any): Promise<void> {
         this.check();
-        await sleep(800);
-        this.self.registerComponent(block.code, block.name, block.namespace, block.suppressRefresh);
-        pluto.third.react.self.requestComponentUpdate();
+        await this.self.registerComponent(block.code, block.name, block.namespace, block.suppressRefresh);
+        await pluto.third.react.self.requestComponentUpdate();
     }
 }